pyproject.toml
[project]
name = "sboxmgr"
version = "1.4.0"
description = "Sing-box config manager CLI (sboxctl)"
authors = [
    { name = "Mikhail Stepanov", email = "kpblcaoo@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "typer>=0.9.0",
    "requests>=2.28.0",
    "python-dotenv",
    "pydantic>=2.0",
]

[project.scripts]
sboxctl = "sboxmgr.cli.main:app"

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-mock",
    "black",
    "pre-commit",
    "detect-secrets",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta" 